feat: Full chatbot features added
Now the chatbot is fully functional. It can answer your queries, it elaborate on financial topics.
A tour has been added for the first time users explain how to use the chatbot.

PDF file support is fully implemented.
When the files are uploading proper messages will be displayed to let the user know about the progress.